added: first files/structure
The present files and project structure were only now considered worth
of versioning. This was so because we decided it was best to keep the
first few versions untracked in order to speed up iteration and because
at that point the project was changing constantly as we experimented with
different structures/designs. Now that the project achieved a bit more stable
structure we can begin version control of incremental changes.