misc: turned ani2d/proc.py into a subpackage
Also added more capabilities to processing derived animations. Now
one can define the order by which a derived animation is processed
and there are also new operations. The "backwards" operation causes
the timing indices to be reversed, so the animation plays backwards
and the "whiten" operation create a copy of a animation where the
visible pixels of the surfaces are completely white (this is good
for making sprites twinkle when they get hit).