add: controls info and message at end of prototype
A message listing controls is now displayed on screen to inform
users.

Another message was added to the intro level at its very end,
encouraging players to consider donating. The message also contains
the URL of online services wherein to donate and also a link to the
game's website so users can know more about it.

Both messages are meant to be temporary, just for the occasion of
the game announcement and release of this first demo/prototype.