change: input playing/recording adjustments
Input playing/recording can already be used when needed.

However, the game itself needs more content/subsystems before
we can begin to use it effectively. This is so not only because
such elements are needed, but because their addition will also
require further changes in input playing/recording (for instance,
when we add support to gamepads/controllers).