add: controls screen and related services
Implemented the controls screen, the panel used to set custom controls.

It can be accessed from the main menu. In the future it will probably
also be available in the pause menu as well (when the user presses the
Escape key during gameplay or the "Start" button on gamepad).

Other services also needed to be created/incremented, in order to
support assignment of custom keys on keyboard and buttons on gamepad,
like incrementing configuration data management code so data describing
the controls could be loaded/saved as needed.

Another service we created was one to automatically manage directional
triggers on the gamepad and generate custom events related to the state
of such triggers, simplifying the usage of such controls a lot. For
instance, the users don't need to configure directional buttons or
other triggers on the gamepad, only other actions, because directionals
are automatically detected and managed.